refactor: apply CR suggestions
