+++ src/core/plugins/json-schema-2020-12-samples/index.js
@@ -18,14 +18,12 @@ import makeGetJsonSampleSchema from "./fn/get-json-sample-schema"
 import makeGetYamlSampleSchema from "./fn/get-yaml-sample-schema"
 import makeGetXmlSampleSchema from "./fn/get-xml-sample-schema"
 import makeGetSampleSchema from "./fn/get-sample-schema"
-import { immutableToJS } from "../../utils"
 
 const JSONSchema202012SamplesPlugin = ({ getSystem }) => {
   const getJsonSampleSchema = makeGetJsonSampleSchema(getSystem)
   const getYamlSampleSchema = makeGetYamlSampleSchema(getSystem)
   const getXmlSampleSchema = makeGetXmlSampleSchema(getSystem)
   const getSampleSchema = makeGetSampleSchema(getSystem)
-  const makeFoldType = (schema) => foldType(immutableToJS(schema)?.type)
 
   return {
     fn: {
@@ -44,7 +42,7 @@ const JSONSchema202012SamplesPlugin = ({ getSystem }) => {
         getXmlSampleSchema,
         getSampleSchema,
         mergeJsonSchema,
-        foldType: makeFoldType,
+        foldType,
       },
     },
   }
