+++ src/core/plugins/json-schema-5-samples/fn/index.js
@@ -628,36 +628,6 @@ export const createXMLExample = (schema, config, o) => {
   return XML(json, { declaration: true, indent: "\t" })
 }
 
-const getType = (schema, processedSchemas = new WeakSet()) => {
-  if (schema == null) {
-    return "any"
-  }
-
-  if (processedSchemas.has(schema)) {
-    return "any" // detect a cycle
-  }
-
-  processedSchemas.add(schema)
-
-  const { type, items } = schema
-
-  const getArrayType = () => {
-    if (items) {
-      const itemsType = getType(items, processedSchemas)
-      return `array<${itemsType}>`
-    } else {
-      return "array<any>"
-    }
-  }
-
-  if (
-    Object.hasOwn(schema, "items")
-  ) {
-    return getArrayType()
-  }
-  return type
-}
-
 
 export const sampleFromSchema = (schema, config, o) =>
   sampleFromSchemaGeneric(schema, config, o, false)
@@ -668,6 +638,4 @@ export const memoizedCreateXMLExample = memoizeN(createXMLExample, resolver)
 
 export const memoizedSampleFromSchema = memoizeN(sampleFromSchema, resolver)
 
-export const getSchemaObjectTypeLabel = (schema) => getType(immutableToJS(schema))
-
-export const getSchemaObjectType = (schema) => schema?.get("type") ?? "string"
+export const getSchemaObjectType = (schema) => immutableToJS(schema)?.type ?? "string"
