+++ src/core/plugins/json-schema-5-samples/index.js
@@ -10,7 +10,6 @@ import {
   memoizedSampleFromSchema,
   mergeJsonSchema,
   getSchemaObjectType,
-  getSchemaObjectTypeLabel,
 } from "./fn/index"
 import makeGetJsonSampleSchema from "./fn/get-json-sample-schema"
 import makeGetYamlSampleSchema from "./fn/get-yaml-sample-schema"
@@ -49,7 +48,6 @@ const JSONSchema5SamplesPlugin = ({ getSystem }) => {
       getXmlSampleSchema,
       getSampleSchema,
       mergeJsonSchema,
-      getSchemaObjectTypeLabel,
       getSchemaObjectType,
     },
   }
