+++ src/core/plugins/json-schema-5/fn.js
@@ -3,6 +3,7 @@
  */
 import { Map } from "immutable"
 import isPlainObject from "lodash/isPlainObject"
+import { immutableToJS } from "core/utils"
 
 export const hasSchemaType = (schema, type) => {
   const isSchemaImmutable = Map.isMap(schema)
@@ -17,3 +18,34 @@ export const hasSchemaType = (schema, type) => {
     type === schemaType || (Array.isArray(type) && type.includes(schemaType))
   )
 }
+
+const getType = (schema, processedSchemas = new WeakSet()) => {
+  if (schema == null) {
+    return "any"
+  }
+
+  if (processedSchemas.has(schema)) {
+    return "any" // detect a cycle
+  }
+
+  processedSchemas.add(schema)
+
+  const { type, items } = schema
+
+  const getArrayType = () => {
+    if (items) {
+      const itemsType = getType(items, processedSchemas)
+      return `array<${itemsType}>`
+    } else {
+      return "array<any>"
+    }
+  }
+
+  if (Object.hasOwn(schema, "items")) {
+    return getArrayType()
+  }
+  return type
+}
+
+export const getSchemaObjectTypeLabel = (schema) =>
+  getType(immutableToJS(schema))
