+++ src/core/plugins/json-schema-5/index.js
@@ -14,7 +14,7 @@ import Schemes from "./components/schemes"
 import SchemesContainer from "./containers/schemes"
 import * as JSONSchemaComponents from "./components/json-schema-components"
 import { ModelExtensions } from "./components/model-extensions"
-import { hasSchemaType } from "./fn"
+import { getSchemaObjectTypeLabel, hasSchemaType } from "./fn"
 
 const JSONSchema5Plugin = () => ({
   components: {
@@ -30,6 +30,7 @@ const JSONSchema5Plugin = () => ({
     ModelExtensions,
     schemes: Schemes,
     SchemesContainer,
+    getSchemaObjectTypeLabel,
     ...JSONSchemaComponents,
   },
   fn: {
