+++ src/core/plugins/oas31/after-load.js
@@ -7,6 +7,7 @@ import {
 } from "./json-schema-2020-12-extensions/fn"
 import { wrapOAS31Fn } from "./fn"
 import { makeIsFileUploadIntended } from "./oas3-extensions/fn"
+import { immutableToJS } from "core/utils"
 
 function afterLoad({ fn, getSystem }) {
   // overrides for fn.jsonSchema202012
@@ -33,8 +34,10 @@ function afterLoad({ fn, getSystem }) {
         getXmlSampleSchema: fn.jsonSchema202012.getXmlSampleSchema,
         getSampleSchema: fn.jsonSchema202012.getSampleSchema,
         mergeJsonSchema: fn.jsonSchema202012.mergeJsonSchema,
-        getSchemaObjectTypeLabel: fn.jsonSchema202012.getType,
-        getSchemaObjectType: fn.jsonSchema202012.foldType,
+        getSchemaObjectTypeLabel: (schema) =>
+          fn.jsonSchema202012.getType(immutableToJS(schema)),
+        getSchemaObjectType: (schema) =>
+          fn.jsonSchema202012.foldType(immutableToJS(schema)?.type),
       },
       getSystem()
     )
